refactor(lsp): improved `HarperToggle` registration implementation
Previously, the `HarperToggle` user command was registered by adding an
entry to the `commands` field of `vim.lsp.Config`. That field is
actually for LSP commands, not user commands related to an LSP. This
changes it to be registered during `on_attach`, in addition to improving
its implementation and fixing some styling issues.